+++ src/Commands/MakeUserCommand.php
@@ -3,6 +3,7 @@
 namespace Spatie\MailcoachUi\Commands;
 
 use Illuminate\Console\Command;
+use Illuminate\Support\Facades\Validator;
 use Spatie\MailcoachUi\Models\User;
 
 class MakeUserCommand extends Command
@@ -18,7 +19,7 @@ class MakeUserCommand extends Command
         $password = $this->option('password');
 
         if (! $username) {
-            $username = $this->ask("What is the username?");
+            $username = $this->ask("What is the user's name?");
         }
 
         if (! $email) {
@@ -27,6 +28,26 @@ class MakeUserCommand extends Command
 
         if (! $password) {
             $password = $this->secret("What is the password?");
+        }
+
+        $validator = Validator::make([
+            'username' => $username,
+            'email' => $email,
+            'password' => $password,
+        ], [
+            'username' => ['required'],
+            'email' => ['required', 'email:rfc,dns', 'unique:users,email'],
+            'password' => ['required', 'min:8'],
+        ]);
+
+        if ($validator->fails()) {
+            $this->info('User not created. See error messages below:');
+
+            foreach ($validator->errors()->all() as $error) {
+                $this->error($error);
+            }
+
+            return 1;
         }
 
         User::create([